Convert float to single precision to reduce memory footprint.

+++ src/tensors.rs
@@ -1,9 +1,9 @@
 pub use crate::utilities::Utilities;
 use ndarray::prelude::*;
 use ndarray_linalg::cholesky::*;
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 
-pub fn lifetime_approx(mut kL: f64) -> f64 {
+pub fn lifetime_approx(mut kL: f32) -> f32 {
     if kL < 0.005 {
         kL = 0.005;
     }
@@ -13,7 +13,7 @@ pub fn lifetime_approx(mut kL: f64) -> f64 {
         / (1.0 - 0.04103886513006046 * kL + 1.1050902034670118 * kSqr)
 }
 
-pub fn vonkarman_spectrum(ae: f64, k: f64, L: f64) -> f64 {
+pub fn vonkarman_spectrum(ae: f32, k: f32, L: f32) -> f32 {
     ae * L.powf(5.0 / 3.0) * (L * k).powi(4) / (1.0 + (L * k).powi(2)).powf(17.0 / 6.0)
 }
 
@@ -30,8 +30,8 @@ pub mod Tensors {
         /// Length scale, $L$
         pub L: T,
     }
-    impl Isotropic<f64> {
-        pub fn from_params(ae: f64, L: f64) -> Isotropic<f64> {
+    impl Isotropic<f32> {
+        pub fn from_params(ae: f32, L: f32) -> Isotropic<f32> {
             Isotropic { ae, L }
         }
     }
@@ -45,21 +45,21 @@ pub mod Tensors {
         pub gamma: T,
     }
 
-    impl Sheared<f64> {
-        pub fn from_params(ae: f64, L: f64, gamma: f64) -> Sheared<f64> {
+    impl Sheared<f32> {
+        pub fn from_params(ae: f32, L: f32, gamma: f32) -> Sheared<f32> {
             Sheared { ae, L, gamma }
         }
 
         /// Isotropic to sheared tensor transformation matrix.
-        pub fn sheared_transform(&self, K: &[f64]) -> Array2<f64> {
-            let k_norm2 = K.iter().map(|x| x * x).sum::<f64>();
+        pub fn sheared_transform(&self, K: &[f32]) -> Array2<f32> {
+            let k_norm2 = K.iter().map(|x| x * x).sum::<f32>();
             if k_norm2 == 0.0 {
                 return Array2::zeros((3, 3));
             }
             let beta = self.gamma * lifetime_approx((k_norm2).sqrt() * self.L);
 
             // Equation (12)
-            let K0: Array1<f64> = arr1(K) + arr1(&[0.0, 0.0, beta * K[0]]);
+            let K0: Array1<f32> = arr1(K) + arr1(&[0.0, 0.0, beta * K[0]]);
             let k0_norm2 = K0.dot(&K0);
             let (zeta1, zeta2);
             if K[0] == 0.0 {
@@ -104,16 +104,16 @@ pub mod Tensors {
         pub min_depth: u64,
     }
 
-    impl ShearedSinc<f64> {
+    impl ShearedSinc<f32> {
         pub fn from_params(
-            ae: f64,
-            L: f64,
-            gamma: f64,
-            Ly: f64,
-            Lz: f64,
-            tol: f64,
+            ae: f32,
+            L: f32,
+            gamma: f32,
+            Ly: f32,
+            Lz: f32,
+            tol: f32,
             min_depth: u64,
-        ) -> ShearedSinc<f64> {
+        ) -> ShearedSinc<f32> {
             ShearedSinc {
                 ae,
                 L,
@@ -126,16 +126,16 @@ pub mod Tensors {
         }
 
         /// Sinc-corrected spectral tensor with additional information.
-        pub fn tensor_info(&self, K: &[f64]) -> (Array2<f64>, u64) {
-            let tensor_gen: Sheared<f64> = Sheared::from_params(self.ae, self.L, self.gamma);
+        pub fn tensor_info(&self, K: &[f32]) -> (Array2<f32>, u64) {
+            let tensor_gen: Sheared<f32> = Sheared::from_params(self.ae, self.L, self.gamma);
 
-            let func = |y: f64, z: f64| {
+            let func = |y: f32, z: f32| {
                 Utilities::sinc2(y * self.Ly / 2.0)
                     * Utilities::sinc2(z * self.Lz / 2.0)
                     * tensor_gen.tensor(&[K[0], K[1] + y, K[2] + z])
             };
 
-            let (mut out, neval): (Array2<f64>, u64) = Utilities::adaptive_quadrature_2d(
+            let (mut out, neval): (Array2<f32>, u64) = Utilities::adaptive_quadrature_2d(
                 func,
                 -2.0 * PI / self.Ly,
                 2.0 * PI / self.Ly,
@@ -158,7 +158,7 @@ pub mod Tensors {
         fn decomp(&self, K: &[T]) -> Array2<T>;
     }
 
-    impl TensorGenerator<f64> for Isotropic<f64> {
+    impl TensorGenerator<f32> for Isotropic<f32> {
         /// Isotropic spectral tensor
         ///
         /// Generates the incompressible isotropic turbulence spectral tensor as
@@ -167,13 +167,13 @@ pub mod Tensors {
         /// \frac{E(|\mathbf{k}|)}{4\pi|\mathbf{k}|^4}(\delta\_{ij}|\mathbf{k}|^2 -
         /// k\_ik\_j) $$
         ///
-        fn tensor(&self, K: &[f64]) -> Array2<f64> {
-            let k_norm = K.iter().map(|x| x * x).sum::<f64>().sqrt();
+        fn tensor(&self, K: &[f32]) -> Array2<f32> {
+            let k_norm = K.iter().map(|x| x * x).sum::<f32>().sqrt();
             if k_norm == 0.0 {
                 return Array2::zeros((3, 3));
             }
             let E = vonkarman_spectrum(self.ae, k_norm, self.L);
-            let mut tensor: Array2<f64> = arr2(&[
+            let mut tensor: Array2<f32> = arr2(&[
                 [K[1].powi(2) + K[2].powi(2), -K[0] * K[1], -K[0] * K[2]],
                 [-K[0] * K[1], K[0].powi(2) + K[2].powi(2), -K[1] * K[2]],
                 [-K[0] * K[2], -K[1] * K[2], K[0].powi(2) + K[1].powi(2)],
@@ -187,31 +187,31 @@ pub mod Tensors {
         /// $\mathbf{\phi}(\mathbf{k})$, where
         /// $\mathbf{\phi}^\*(\mathbf{k})\mathbf{\phi}(\mathbf{k}) =
         /// \mathbf{\Phi}^{\text{ISO}}(\mathbf{k})$.
-        fn decomp(&self, K: &[f64]) -> Array2<f64> {
-            let k_norm = K.iter().map(|x| x * x).sum::<f64>().sqrt();
+        fn decomp(&self, K: &[f32]) -> Array2<f32> {
+            let k_norm = K.iter().map(|x| x * x).sum::<f32>().sqrt();
             if k_norm == 0.0 {
                 return Array2::zeros((3, 3));
             }
             let E = vonkarman_spectrum(self.ae, k_norm, self.L);
-            let mut tensor: Array2<f64> =
+            let mut tensor: Array2<f32> =
                 arr2(&[[0.0, K[2], -K[1]], [-K[2], 0.0, K[0]], [K[1], -K[0], 0.0]]);
             tensor *= (E / PI).sqrt() / (2.0 * k_norm.powi(2));
             tensor
         }
     }
 
-    impl TensorGenerator<f64> for Sheared<f64> {
+    impl TensorGenerator<f32> for Sheared<f32> {
         /// Sheared (Mann) spectral tensor
-        fn tensor(&self, K: &[f64]) -> Array2<f64> {
-            let k_norm2 = K.iter().map(|x| x * x).sum::<f64>();
+        fn tensor(&self, K: &[f32]) -> Array2<f32> {
+            let k_norm2 = K.iter().map(|x| x * x).sum::<f32>();
             if k_norm2 == 0.0 {
                 return Array2::zeros((3, 3));
             }
-            let A: Array2<f64> = self.sheared_transform(K);
+            let A: Array2<f32> = self.sheared_transform(K);
             let beta = self.gamma * lifetime_approx((k_norm2).sqrt() * self.L);
 
             // Equation (12)
-            let K0: Array1<f64> = arr1(K) + arr1(&[0.0, 0.0, beta * K[0]]);
+            let K0: Array1<f32> = arr1(K) + arr1(&[0.0, 0.0, beta * K[0]]);
             let iso_tensor = Isotropic {
                 ae: self.ae,
                 L: self.L,
@@ -221,16 +221,16 @@ pub mod Tensors {
             A.dot(&iso_tensor).dot(&A.t())
         }
         /// Decomposition of sheared (Mann) spectral tensor
-        fn decomp(&self, K: &[f64]) -> Array2<f64> {
-            let k_norm2 = K.iter().map(|x| x * x).sum::<f64>();
+        fn decomp(&self, K: &[f32]) -> Array2<f32> {
+            let k_norm2 = K.iter().map(|x| x * x).sum::<f32>();
             if k_norm2 == 0.0 {
                 return Array2::zeros((3, 3));
             }
-            let A: Array2<f64> = self.sheared_transform(K);
+            let A: Array2<f32> = self.sheared_transform(K);
             let beta = self.gamma * lifetime_approx((k_norm2).sqrt() * self.L);
 
             // Equation (12)
-            let K0: Array1<f64> = arr1(K) + arr1(&[0.0, 0.0, beta * K[0]]);
+            let K0: Array1<f32> = arr1(K) + arr1(&[0.0, 0.0, beta * K[0]]);
             let iso_tensor = Isotropic {
                 ae: self.ae,
                 L: self.L,
@@ -241,16 +241,16 @@ pub mod Tensors {
         }
     }
 
-    impl TensorGenerator<f64> for ShearedSinc<f64> {
+    impl TensorGenerator<f32> for ShearedSinc<f32> {
         /// Sheared spectral tensor with sinc correction
 
-        fn tensor(&self, K: &[f64]) -> Array2<f64> {
-            let (out, _): (Array2<f64>, u64) = self.tensor_info(K);
+        fn tensor(&self, K: &[f32]) -> Array2<f32> {
+            let (out, _): (Array2<f32>, u64) = self.tensor_info(K);
             out
         }
 
         /// Decomposition of sheared spectral tensor with sinc correction
-        fn decomp(&self, K: &[f64]) -> Array2<f64> {
+        fn decomp(&self, K: &[f32]) -> Array2<f32> {
             self.tensor(K).cholesky(UPLO::Lower).unwrap()
         }
     }
